pyproject.toml


[project]
name = "telegram-bookmark-bot"
version = "0.1.0"
description = "AI-powered Telegram bot for message summarization, classification, and storage to Notion/Obsidian"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Yan Lu", email = "luyanfcp@gmail.com"},
]
keywords = ["telegram", "bot", "ai", "notion", "obsidian"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Core
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    # Storage
    "notion-client>=2.0.0",
    "python-frontmatter>=1.1.0",
    # Content extraction
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.1",
    "pydantic>=2.5.0",
    "pillow>=10.1.0",
    "pytesseract>=0.3.10",
    "youtube-transcript-api>=0.6.2",
    # Infrastructure
    "redis>=5.0.0",
    "celery>=5.3.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "structlog>=23.2.0",
    "sentry-sdk>=1.39.0",
    # Configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    # Optional: For Word document processing
    # "python-docx>=1.1.0",
    "anthropic>=0.75.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/telegram-knowledge-bot"
Repository = "https://github.com/yourusername/telegram-knowledge-bot"
Issues = "https://github.com/yourusername/telegram-knowledge-bot/issues"

[project.scripts]
telegbot = "src.run:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
